+++ references/segmentation/v2_extras.py
@@ -11,7 +11,7 @@ class PadIfSmaller(v2.Transform):
         self.fill = v2._utils._setup_fill_arg(fill)
 
     def _get_params(self, sample):
-        _, height, width = v2.utils.query_chw(sample)
+        _, height, width = v2._utils.query_chw(sample)
         padding = [0, 0, max(self.size - width, 0), max(self.size - height, 0)]
         needs_padding = any(padding)
         return dict(padding=padding, needs_padding=needs_padding)
