+++ torchvision/transforms/v2/_color.py
@@ -6,7 +6,7 @@ from torchvision import transforms as _transforms
 from torchvision.transforms.v2 import functional as F, Transform
 
 from ._transform import _RandomApplyTransform
-from .utils import query_chw
+from ._utils import query_chw
 
 
 class Grayscale(Transform):
